Weight overflow fix, thanks jiv for idea.
git-svn-id: http://svn.l2jserver.com/trunk@3377 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/model/actor/instance/L2PcInstance.java
@@ -2095,8 +2095,8 @@ public final class L2PcInstance extends L2Playable
 		int maxLoad = getMaxLoad();
 		if (maxLoad > 0)
 		{
-			int weightproc = getCurrentLoad() * 1000 / maxLoad;
-			weightproc*=(int)calcStat(Stats.WEIGHT_LIMIT, 1, this, null);
+			long weightproc = (long)getCurrentLoad() * 1000 / maxLoad;
+			weightproc *= calcStat(Stats.WEIGHT_LIMIT, 1, this, null);
 			int newWeightPenalty;
 			if (weightproc < 500 || _dietMode)
 			{
